+++ src/components/sections/elements/GeographicExtent.vue
@@ -4,7 +4,6 @@ import { computed, type ComputedRef, nextTick, onMounted, ref, watch } from 'vue
 import { Stability, SectionType } from '@/types/enum'
 import { getExtents, getExtent, getProjection } from '@/lib/data'
 import { createExtent, createProjection } from '@/lib/extents'
-import type { EsriToken, WellKnownExtent } from '@/types/app'
 import type { Extent, ReferenceSystemInfo } from '@/types/iso'
 
 import SectionBorder from '@/components/bases/SectionBorder.vue'
@@ -13,7 +12,6 @@ import Output from '@/components/bases/Output.vue'
 import FormLabel from '@/components/bases/FormLabel.vue'
 import TwoColumn from '@/components/bases/TwoColumn.vue'
 import GeographicExtentMap from '@/components/sections/elements/GeographicExtentMap.vue'
-import GeographicExtentGlobe from '@/components/sections/elements/GeographicExtentGlobe.vue'
 
 defineProps({
   esriToken: {
@@ -110,12 +108,6 @@ watch(extent, () => {
         </template>
         <template #right>
           <div class="space-y-2">
-            <div class="text-sky-500">Preview (3D)</div>
-            <GeographicExtentGlobe
-              v-if="renderMaps"
-              :wke="wellKnownExtent"
-              :esri-token="esriToken"
-            />
           </div>
         </template>
       </TwoColumn>