+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 * Research topics component
 * Esri identifier (as not a 1:1 mapping)
 * WMS service (as no longer supported, replaced with placeholder to ensure services section still works)
+* 3D spatial extent visualisation
 
 ## [0.10.4] - 2024-10-18